moved useful classes out of 'entities' folder
