+++ src/main/java/com/marginallyclever/robotoverlord/ViewCube.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.robotoverlord.entities;
+package com.marginallyclever.robotoverlord;
 
 import com.jogamp.opengl.GL2;
 import com.marginallyclever.convenience.MatrixHelper;
